fix wait dialog state
