+++ tests/copyFolderButton.test.js
@@ -10,14 +10,20 @@ function loadMain() {
     global.navigator.clipboard = { writeText: jest.fn().mockResolvedValue() };
     global.updateStatus = jest.fn();
     global.window.electronAPI = {};
-    ({ copyDownloadFolder, showDownloadWaitDialog, __setFiles } = require('../web/src/main.js'));
+    global.files = [];
+    global.escapeHtml = t => t;
+    global.safeCopy = async text => { await navigator.clipboard.writeText(text); return true; };
+    ({ copyDownloadFolder, showDownloadWaitDialog } = require('../web/src/dubbing.js'));
+    ({ __setFiles } = require('../web/src/main.js'));
 }
 
 describe('copyDownloadFolder', () => {
     beforeEach(loadMain);
 
     test('kopiert den Ordner erneut', async () => {
-        __setFiles([{ id: 1, folder: 'TestOrdner', enText: '', deText: '' }]);
+        const arr = [{ id: 1, folder: 'TestOrdner', enText: '', deText: '' }];
+        __setFiles(arr);
+        global.files = arr;
         document.body.innerHTML = '';
         await showDownloadWaitDialog(1);
         navigator.clipboard.writeText.mockClear();
@@ -26,7 +32,9 @@ describe('copyDownloadFolder', () => {
     });
 
     test('Button im Dialog vorhanden', async () => {
-        __setFiles([{ id: 2, folder: 'OrdnerB', enText: '', deText: '' }]);
+        const arr2 = [{ id: 2, folder: 'OrdnerB', enText: '', deText: '' }];
+        __setFiles(arr2);
+        global.files = arr2;
         document.body.innerHTML = '';
         await showDownloadWaitDialog(2);
         const btn = document.getElementById('copyFolderBtn');
